+++ src/Http/Controllers/Controller.php
@@ -48,7 +48,7 @@ class Controller extends BaseController
     /**
      * The class name of Resource to be used for the show and index methods.
      *
-     * @var string $resource
+     * @var string
      */
     protected $resource = BaseResource::class;
 
@@ -68,6 +68,7 @@ class Controller extends BaseController
 
     /**
      * @param null $model
+     *
      * @throws Exception
      */
     public function __construct($model = null)