+++ src/Http/Controllers/Filters/WhereIn.php
@@ -13,6 +13,7 @@ class WhereIn extends Filter
     /**
      * @param Controller $controller
      * @param Builder|Relation $query
+     *
      * @throws Exception
      */
     public static function apply(Controller $controller, Builder|Relation $query): void